refactor(axi): width-derived strobe type and size limits, value-preserving at 256 b
Widen WBeat/MemWriteReq/wrap-IO strb fields to uint64_t across the 6 hardcoded
uint32_t sites (types.hpp, memory_port.hpp, protocol_rules.hpp, scoreboard.hpp,
axi_master.hpp, both *_wrap_io.hpp), plus AxiMaster's strb_per_beat vectors and
load_strb_file_, and a truncating cast in nsu/depacketize.hpp found by sweep.

protocol_rules::kMaxSize is now derived (log2_pow2(DATA_BYTES)) instead of a
literal 5. check_strb_valid_bits drops the `if constexpr (DATA_BYTES >= 32)
return true` shortcut that silently disabled the check at wider widths, in
favor of a real width-derived mask (new axi::kFullStrbMask, shared with
load_strb_file_'s default-full value).

DATA_BYTES stays 32; nothing flips. New death test
StrbValidBits_RejectsBitAboveDataBytes guards against the removed shortcut.
506/506 ctest, 0 regressions.

+++ src/c_model/include/axi/memory_port.hpp
@@ -7,7 +7,7 @@ namespace ni::cmodel::axi {
 struct MemWriteReq {
     uint64_t addr;
     std::array<uint8_t, DATA_BYTES> data;
-    uint32_t strb;
+    uint64_t strb;
     uint8_t id;
     bool last;
     uint64_t tag;
